cleanup cluster builder
