+++ pilot/pkg/networking/core/v1alpha3/cluster_waypoint.go
@@ -91,7 +91,7 @@ func (cb *ClusterBuilder) buildWaypointInboundVIPCluster(svc *model.Service, por
 	clusterName := model.BuildSubsetKey(model.TrafficDirectionInboundVIP, subset, svc.Hostname, port.Port)
 
 	clusterType := cluster.Cluster_EDS
-	localCluster := cb.buildDefaultCluster(clusterName, clusterType, nil,
+	localCluster := cb.buildCluster(clusterName, clusterType, nil,
 		model.TrafficDirectionInbound, &port, nil, nil)
 
 	// Ensure VIP cluster has services metadata for stats filter usage
